backend/requirements.txt
python-dotenv
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
certifi            
charset-normalizer
contourpy          
cycler            
filelock           
fonttools          
fsspec             
idna               
Jinja2            
joblib            
kiwisolver         
MarkupSafe         
matplotlib        
mpmath            
networkx          
numpy            
packaging       
pandas         
pillow           
pip                
pyparsing         
python-dateutil    
python-dotenv     
pytz               
requests           
scikit-learn       
scikit-surprise   
scipy             
six               
surprise          
sympy              
threadpoolctl      
torch   
tqdm           
typing_extensions  
tzdata             
urllib3           